// server/src/api.d.ts
export interface SerieResume {
	id: string;
	name: string;
	logo?: string;
}

/**
 * /series/:id endpoint
 */
export interface Serie extends SerieResume {
	firstSet: SetResume;
	lastSet: SetResume;
	releaseDate: string;
	sets: Array<SetResume>;
}

interface variants {
	normal?: boolean;
	reverse?: boolean;
	holo?: boolean;
	firstEdition?: boolean;
	wPromo?: boolean
}

interface variant_detailed {
	type: string
	size?: string
	stamp?: Array<string>
	foil?: string
}

export interface SetResume {
	id: string;
	name: string;
	logo?: string;
	symbol?: string;
	cardCount: {
		/**
		 * total of number of cards
		 */
		total: number;
		/**
		 * number of cards officialy (on the bottom of each cards)
		 */
		official: number;
	};
}